dashboard renamed to home/ home renamed to artist
